fix doctests

+++ src/snowflake/snowpark/modin/plugin/docstrings/series.py
@@ -2995,7 +2995,7 @@ class Series:  # pragma: no cover: we use this class's docstrings, but we never 
         2    5
         3    7
         4    7
-        dtype: int8
+        dtype: int64
 
         >>> s.nunique()
         4
